requirements.txt
asgiref==3.8.1
click==8.1.8
colorama==0.4.6
cssbeautifier==1.15.3
Django==5.1.6
django-auditlog==3.0.0
django-bootstrap5==24.3
djlint==1.36.4
EditorConfig==0.17.0
jsbeautifier==1.15.3
json5==0.10.0
pathspec==0.12.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
ruff==0.9.7
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
